feat: adding create_lobby tests and lobby name validation tests
small refactors as well to resolve warnings from cargo clippy, removed sanitize() function as it was unused and improper

+++ src/web/game/player.rs
@@ -64,13 +64,13 @@ where
     }
 
     pub fn set_wager(&mut self, wager: i32) -> Result<(), PlayerError<T>> {
-        if wager as i32 > self.points || wager < 0 {
+        if wager > self.points || wager < 0 {
             return Err(PlayerError::User(UserError::InvalidWager(
                 wager,
                 self.points,
             )));
         }
-        self.wager = wager as i32;
+        self.wager = wager;
         Ok(())
     }
